Added grades, subject, expanding article. Played around with some new styles.
