+++ next.config.js
@@ -1,15 +1,14 @@
 // /** @type {import('next').NextConfig} */
 
-
 module.exports = {
   images: {
     remotePatterns: [
       {
-        protocol: 'https',
-        hostname: 'github.com',
+        protocol: "https",
+        hostname: "github.com",
         // port: '443',
         // pathname: '/*'
       },
     ],
   },
-}
+};
