+++ src/styles/Form.module.css
@@ -1,39 +1,36 @@
 /* form element style */
-.loginForm, .loginLabel {
-    margin: 0.5rem 1rem 1rem;
-    display: flex;
-    flex-flow: column;
+.loginForm,
+.loginLabel {
+  margin: 0.5rem 1rem 1rem;
+  display: flex;
+  flex-flow: column;
 }
 
 .formLayout {
-    min-width: 100%;
-    min-height: 100%;
+  min-width: 100%;
+  min-height: 100%;
 }
 
 .loginPrompt {
-    font-weight: 600;
+  font-weight: 600;
 }
 
 .formInput {
-    padding: 8px;
-    margin: 0.5rem 0 1rem;
-    border: 1px solid #ccc;
-    border-radius: 4px;
+  padding: 8px;
+  margin: 0.5rem 0 1rem;
+  border: 1px solid #ccc;
+  border-radius: 4px;
 }
 
 .loginTitle {
-    margin: 0.5rem 1rem 1rem;
-    font-family: Arial, Helvetica, sans-serif;
+  margin: 0.5rem 1rem 1rem;
 }
 
 .loginSubmit {
-    margin: 0.5rem 1rem 1rem;
+  margin: 0.5rem 1rem 1rem;
 }
 
-
-
-
 .error {
-    color: brown;
-    margin: 1rem 0 0;
+  color: brown;
+  margin: 1rem 0 0;
 }
