+++ src/styles/List.module.css
@@ -1,16 +1,20 @@
 .mainContainer {
-  background: #343434;
+  cursor: url("https://storage.googleapis.com/scr-notes/licensed-image.jpeg"),
+    auto;
+  /* background: #343434; */
   padding: 10px 20vw 10px 10vw;
 }
 
 .lessonContainer {
-  background: #fff;
+  background: #ffffff;
+  color: black;
   padding: 25px;
   margin: 25px 0;
   border-radius: 5px;
   max-width: 80vw;
   margin: auto auto;
   margin-top: 20px;
+  box-shadow: #1c1c1c81 4px 4px 2px;
 }
 
 .authorText {
@@ -36,6 +40,27 @@
   color: #b7b4c4;
 }
 
-.divider {
-  margin: 0;
+.lessonAbstract {
+  /* line-clamp 
+   * https://stackoverflow.com/questions/3922739/limit-text-length-to-n-lines-using-css/13924997#13924997 
+  */
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  width: 80;
+}
+
+.lessonAbstract:hover {
+  cursor: text;
+  background-color: rgba(0, 0, 0, 0.08);
+  border-radius: 4px;
+}
+
+.lessonAbstractCollapsed {
+  -webkit-line-clamp: 3; /* number of lines to show */
+  line-clamp: 3;
+}
+
+.gradeBox {
+  margin-top: 10px;
 }
