+++ src/styles/globals.css
@@ -1,9 +1,9 @@
 :root {
   --max-width: 1100px;
   --border-radius: 12px;
-  --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
-    'Roboto Mono', 'Oxygen Mono', 'Ubuntu Monospace', 'Source Code Pro',
-    'Fira Mono', 'Droid Sans Mono', 'Courier New', monospace;
+  --font-mono: ui-monospace, Menlo, Monaco, "Cascadia Mono", "Segoe UI Mono",
+    "Roboto Mono", "Oxygen Mono", "Ubuntu Monospace", "Source Code Pro",
+    "Fira Mono", "Droid Sans Mono", "Courier New", monospace;
 
   --foreground-rgb: 0, 0, 0;
   --background-start-rgb: 214, 219, 220;
@@ -46,7 +46,10 @@
     --background-start-rgb: 0, 0, 0;
     --background-end-rgb: 0, 0, 0;
 
-    --primary-glow: radial-gradient(rgba(1, 65, 255, 0.24), rgba(1, 65, 255, 0));
+    --primary-glow: radial-gradient(
+      rgba(1, 65, 255, 0.24),
+      rgba(1, 65, 255, 0)
+    );
     --secondary-glow: linear-gradient(
       to bottom right,
       rgba(1, 65, 255, 0),
@@ -87,11 +90,14 @@ body {
 
 body {
   color: rgb(var(--foreground-rgb));
+
+  background: rgb(2, 0, 36);
   background: linear-gradient(
-      to bottom,
-      rgb(var(--background-end-rgb))
-    )
-    rgb(var(--background-start-rgb));
+    70deg,
+    rgba(2, 0, 36, 1) 0%,
+    rgba(104, 34, 139, 1) 100%,
+    rgba(0, 144, 255, 1) 100%
+  );
 }
 
 a {
